requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
fonttools==4.54.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Locales==0.0.2
logging==0.4.9.6
MarkupSafe==2.1.5
matplotlib==3.9.2
MouseInfo==0.1.3
mysql-connector-python==9.0.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pywhatkit==5.4
reportlab==4.2.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
types-Flask-Cors==4.0.0.20240806
types-PyAutoGUI==0.9.3.20240406
types-PyScreeze==1.0.1.20240822
types-reportlab==4.2.2.20240819
Unidecode==1.3.8
urllib3==2.2.2
Werkzeug==3.0.3
wikipedia==1.4.0
